requirements.txt
dotenv
autogen-ext[ollama]
ag2[ollama]
openai
ag2[openai]
pyautogen
streamlit

